// projects/num-lib/src/lib/shared/models/user/user-profile.interface.ts
/**
 * Copyright 2021 Vitagroup AG
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { AvailableRoles } from '../available-roles.enum'
import { IOrganization } from '../organization/organization.interface'

export interface IUserProfile {
  id: string
  username: string
  firstName: string
  lastName: string
  email: string
  createdTimestamp: number
  roles?: AvailableRoles[]
  approved: boolean
  organization: IOrganization
}